refactor: 탭 뷰 컴포넌트 이름 변경
