--- src/components/BakeryDetail/Report/InformationEditView/index.ts
@@ -1,1 +1,0 @@
-export * from './InformationEditView';
