+++ src/components/BakeryDetail/Report/InformationReportView/index.ts
@@ -1,0 +1,1 @@
+export * from './InformationReportView';
