--- src/components/BakeryDetail/Report/MenuEditView/index.ts
@@ -1,1 +1,0 @@
-export * from './MenuEditView';
