+++ src/components/BakeryDetail/Report/MenuReportView/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuReportView';
